exercise week 3/PetElevator/HR/Employee.cs
using System.Collections.Generic;
using PetElevator.Shared;

namespace PetElevator.HR
{
    public class Employee : Person, IBillable
    {
        public int EmployeeId { get; private set; }
        public string Title { get; set; }
        public Department Department { get; set; }
        public double Salary { get; set; }

        public override string FullName
        {
            get
            {
                return $"{LastName}, {FirstName}";
            }
        }

        public Employee(string firstName, string lastName) : this(firstName, lastName, "", 0)
        {
        }

        public Employee(string firstName, string lastName, string title, double salary) : base(firstName, lastName)
        {
            Title = title;
            Salary = salary;
        }

        public void RaiseSalary(double percentage)
        {
            if (percentage > 0)
            {
                Salary *= (percentage / 100) + 1;
            }
        }
        
        public double GetBalanceDue(Dictionary<string, double> servicesRendered)
        {
            foreach (KeyValuePair<string, double> service in servicesRendered)
            {
                if (servicesRendered.ContainsKey("Walking"))
                {
                    return service.Value / 2;
                }
                return service.Value;
            }
            return 0.00;
        }

    }
}
